+++ src/sentry/search/eap/utils.py
@@ -1,7 +1,19 @@
 from datetime import datetime
+from typing import Any, Callable
 
 from google.protobuf.timestamp_pb2 import Timestamp
 from sentry_protos.snuba.v1.endpoint_time_series_pb2 import TimeSeriesRequest
+
+from sentry.exceptions import InvalidSearchQuery
+
+
+def literal_validator(values: list[Any]) -> Callable[[str], bool]:
+    def _validator(input: str) -> bool:
+        if input in values:
+            return True
+        raise InvalidSearchQuery(f"Invalid parameter {input}. Must be one of {values}")
+
+    return _validator
 
 
 def add_start_end_conditions(
